test: move test to traceevent_test package
